+++ src/Renderer/Fingerprint.php
@@ -35,6 +35,7 @@ final class Fingerprint
             'campaignId' => $banner->getCampaignId(),
             'campaignCode' => $banner->getCampaignCode(),
             'campaignName' => $banner->getCampaignName(),
+            'closeExpiration' => $banner->getCloseExpiration(),
         ];
 
         try {